# Conflicts:
#	.documentation/Frontend_to_Dash_Api.yml

+++ django_server/farminsight_dashboard_backend/services/membership_services.py
@@ -18,34 +18,45 @@ def create_membership(data: dict) -> MembershipSerializer:
     return membership_serializer
 
 
-def update_membership(membership_id, new_membership_role):
+def update_membership(membership_id: str, new_membership_role: str, requesting_user: Userprofile) -> MembershipSerializer:
     """
     An Admin of the Organization, or System Admin of the Backend can promote a user.
     :param membership_id:
     :param new_membership_role:
-    :return:
+    :param requesting_user:
+    :return: success
     """
+    try:
+        membership = Membership.objects.select_related('userprofile').get(id=membership_id)
+    except Membership.DoesNotExist:
+        raise NotFoundException(f'Membership {membership_id} not found.')
+
+    if new_membership_role == MembershipRole.Member:
+        can_update = False
+        if is_system_admin(requesting_user) or not membership.userprofile.is_active:
+            can_update = True
+    else:
+        can_update = True
+
+    if can_update:
+        membership.membershipRole = new_membership_role
+        membership.save()
+        return MembershipSerializer(membership)
+
+    raise PermissionDenied()
+
+
+def remove_membership(membership_id) -> bool:
     try:
         membership = Membership.objects.get(id=membership_id)
     except Membership.DoesNotExist:
         raise NotFoundException(f'Membership {membership_id} not found.')
 
-    membership.membershipRole = new_membership_role
-    membership.save()
-
-
-def remove_membership(membership_id):
-    """
-    Only an admin can delete a user.
-    :param membership_id:
-    :return:
-    """
-    try:
-        membership = Membership.objects.get(id=membership_id)
-    except Membership.DoesNotExist:
-        raise NotFoundException(f'Membership {membership_id} not found.')
+    if membership.membershipRole == MembershipRole.Admin:
+        return False
 
     membership.delete()
+    return True
 
 
 def is_member(user, organization: Organization):
@@ -66,6 +77,7 @@ def is_admin(user, organization: Organization):
 
 def is_system_admin(user):
     return getattr(user, "systemRole", None) == SystemRole.SystemAdmin.value
+
 
 def get_memberships_by_organization(organization_id):
     return Membership.objects.filter(organization_id=organization_id)
